Rename LibraryLogic to LibraryStorage and refactor LibraryLogic to under Storage
