+++ src/main/java/seedu/address/model/library/Library.java
@@ -9,6 +9,7 @@ import javafx.collections.ObservableList;
 import seedu.address.model.ReadOnlyLibrary;
 import seedu.address.model.book.Book;
 import seedu.address.model.person.Person;
+import seedu.address.storage.LibraryStorage;
 
 /**
  * Represents a collection of books in a library.
@@ -27,7 +28,7 @@ public class Library implements ReadOnlyLibrary {
 
     private ObservableList<Book> bookList;
     private Threshold threshold;
-    private LibraryLogic libraryLogic;
+    private LibraryStorage libraryLogic;
 
     /**
      * Construct an empty library.
@@ -35,7 +36,7 @@ public class Library implements ReadOnlyLibrary {
     public Library() {
         bookList = FXCollections.observableArrayList();
         threshold = new Threshold();
-        libraryLogic = new LibraryLogic();
+        libraryLogic = new LibraryStorage();
     }
 
     /**
